Add kwinputs to Kineto Traces (#130373)
Summary:
Pull Request resolved: https://github.com/pytorch/pytorch/pull/130373

On the autograd side of things, we are currently saving the kwinputs but we aren't doing anything with them on the profiler side. This diff enables the use of the kwinputs for both FunctionEvents and Chrome Traces.

Test Plan: Added unit testing for both chrome traces and FunctionEvents. Used RecordFunctionFast to test kwinputs since test already had kwargs being passed in but not tested.

Reviewed By: davidberard98

Differential Revision: D59472345